% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_helpers.R
\name{av_size}
\alias{av_size}
\title{Average tree size in forest}
\usage{
av_size(forest)
}
\arguments{
\item{forest}{A splineforest object}
}
\value{
avergae number of terminal nodes
}
\description{
Returns the average number of terminal nodes for tree in forest
}
\examples{
data(sample_forest)
av_size(sample_forest)
}
