% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_helpers.R
\name{avSize}
\alias{avSize}
\title{Average tree size in forest}
\usage{
avSize(forest)
}
\arguments{
\item{forest}{A splineforest object}
}
\value{
avergae number of terminal nodes
}
\description{
Returns the average number of terminal nodes for tree in forest
}
\examples{
avSize(sample_forest)
}
