% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineTree.R
\name{flatten_predictors}
\alias{flatten_predictors}
\title{Flattens predictor variable data into one row per person}
\usage{
flatten_predictors(idvar, data)
}
\arguments{
\item{idvar}{The string name of the ID variable (used to group observations)}

\item{data}{The full dataset to be flattened (long form)}
}
\value{
A wide format dataset with spline coefficients as the responses.
}
\description{
Assumes that splitting explanatory variables do not vary with time. Spline Tree is not meant to handle time-varying covariates.
}
\examples{
flatten_predictors('ID', nlsySample)
}
\keyword{internal}
