% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forest}
\alias{forest}
\title{Sample forest used in vignettes}
\format{An object of class \code{list} of length 15.}
\usage{
forest
}
\description{
Sample forest used in vignettes
}
\keyword{internal}
