% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_useful_fns.R
\name{getNodeData}
\alias{getNodeData}
\title{Returns the portion of the data found at a given terminal node}
\usage{
getNodeData(tree, node, dataType = "all")
}
\arguments{
\item{tree}{a splinetree object}

\item{node}{The number of the node that you want the data for.
Node numbers for your model can be seen using stPrint(tree)
or treeSummary(tree). Note that this node number should correspond to
a terminal node.}

\item{dataType}{If "all", the data returned is the original data (one row per individual observation
with original response values). If "flat", the data returned is the flattened data (one row per person/unit),
with spline coefficients instead of response values.}
}
\value{
A dataframe which holds all the data that falls into this node of the tree.
}
\description{
Given a terminal node number, this function returns the dataset found at this terminal node.
If the dataType arguement is 'all', then all rows of data (with original response values) that
fall in this node are returned.  Otherwise, the flattened data is returned (one row of data per
person/unit, original responses replaced by spline coefficients).
}
