% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{importance}
\alias{importance}
\title{Sample importance used in vignettes}
\format{An object of class \code{list} of length 3.}
\usage{
importance
}
\description{
Sample importance used in vignettes
}
\keyword{internal}
