% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineTree.R
\name{individual_spline}
\alias{individual_spline}
\title{Get spline coefficients for a single person}
\usage{
individual_spline(person, idvar, yvar, tvar, data, boundaryKnots,
  innerKnots, degree, intercept)
}
\arguments{
\item{person}{ID of this person}

\item{idvar}{name of the id variable (string)}

\item{yvar}{the name of the response variable}

\item{tvar}{name of time variable (string)}

\item{data}{full dataset}

\item{boundaryKnots}{the boundary knots for the bspline}

\item{innerKnots}{the inner knots for the bspline}

\item{degree}{the degree of the bspline}

\item{intercept}{whether or not to include an intercept}
}
\description{
Get spline coefficients for a single person
}
\keyword{internal}
