% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlsySample_large}
\alias{nlsySample_large}
\title{Additional socioeconomic information and BMIs of 1000 individuals.}
\format{A data frame with 8083 rows and 18 columns.
\describe{
  \item{ID}{jj}
  \item{News}{gj}
  \item{SEX}{gnegne}
  \item{AGE}{aa}
  \item{BLACK}{aa}
  \item{BMI}{aa}
  \item{Father_14}{aa}
  \item{HEIGHTIN}{aa}
  \item{HGC_FATHER}{aa}
  \item{HGC_MOTHER}{aa}
  \item{HISP}{aa}
  \item{Lib}{aa}
  \item{Mag}{aa}
  \item{HGC}
  \item{Mother_14}{aa}
  \item{Num_sibs}{aa}
  \item{Two_Adults_14}{aa}
  \item{WEIGHT}{aa}
  \item{WHITE}{aa}
  \item{Dad_Full_Work}{}
  \item{Mom_Full_Work}{}
  \item{Age_first_alc}{}
  \item{POVSTAT}{}
  \item{Age_first_smoke}{}
  \item{STABLE_RESIDENCE}{}
  \item{Live_with_parents}{}
  \item{Age_first_weed}{}
  \item{Mother_on_record}{}
  \item{Father_on_record}{}
  \item{Mom_dad_household}{}
  \item{South_Birth}{}
  \item{URBAN_14}{}
  \item{HOUSEHOLD_14}{}
  \item{RACE}{}
}}
\source{
\url{https://www.bls.gov/nls/nlsy79.htm}
}
\usage{
nlsySample_large
}
\description{
See the description for the nlsySample data. This is an extended version of that dataset
with 1000 individuals drawn from the same population of survey respondents. There are also
covariates in this dataset that were excluded from the smaller version of the dataset.
}
\keyword{datasets}
