% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{nodePlot}
\alias{nodePlot}
\title{Plots the trajectories of each terminal node side by side.}
\usage{
nodePlot(model, colors = NULL)
}
\arguments{
\item{model}{A model returned from splineTree()}

\item{colors}{A list of colors to use. By default, uses colors drawn from a rainbow.}
}
\description{
Corresponds to plotting only the second panel of stPlot(). If model$intercept==FALSE, estimated
intercepts are added to each trajectory so that the trajectories are plotted at the level of reasonable
response values.
}
