% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance_plotting.R
\name{plot_varimp}
\alias{plot_varimp}
\title{Create a variable importance plot}
\usage{
plot_varimp(importance_vector)
}
\arguments{
\item{importance_vector}{a named vector where the names are the variables and the vector stores the importances.}
}
\description{
Pass in a named vector of variable importance measures. This function will make a barplot of the importances. The importances are scaled, so
only relative importance is shown.
}
