% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_useful_fns.R
\name{plotNodeTraj}
\alias{plotNodeTraj}
\title{Plot the predicted trajectory for a single node}
\usage{
plotNodeTraj(tree, node, includeData = FALSE)
}
\arguments{
\item{tree}{a splinetree object}

\item{node}{a node number that must correspond to a terminal node}

\item{includeData}{would you like to see the data from the node
plotted along with the predicted trajectory?}
}
\description{
Make a simple plot to view the trajectory predicted at a given node. Option to include or not include
the individual trajectories of people in the node as well.
}
