% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_prediction.R
\name{predict_coeffs_tree}
\alias{predict_coeffs_tree}
\title{Predict spline coefficients for a testset using a single tree.}
\usage{
predict_coeffs_tree(tree, testset = tree$data)
}
\arguments{
\item{tree}{a splinetree object}

\item{testset}{the dataset to return predictions for. If omitted, defaults
to the data used to build this tree.}
}
\value{
a matrix of predicted coefficients
}
\description{
Predict spline coefficients for a testset using a single tree.
}
