% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_evaluation.R
\name{predict_spline_coeffs}
\alias{predict_spline_coeffs}
\title{Predict spline coefficients from a splinetree object}
\usage{
predict_spline_coeffs(tree, testset = tree$parms$flat_data)
}
\arguments{
\item{tree}{A splinetree object}

\item{testset}{The dataset to predict coefficients for. Default is the dataset used to make the tree.}
}
\value{
A matrix of spline coefficients. Dimension is number of units in test set by degrees of freedom of the spline.
}
\description{
Returns a matrix of spline coefficients for each observation in the testset. If no testset is provided,
returns predicted coefficients for the individuals in the dataset used to build the tree.
}
\details{
importFrom treeClust rpart.predict.leaves
}
\examples{
splitForm <- BMI~HISP+WHITE+BLACK+HGC_MOTHER+SEX
tree <- splineTree(splitForm, BMI~AGE, "ID", nlsySample, degree=1, intercept=FALSE, cp=0.005)
predict_spline_coeffs(tree)
}
