% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_evaluation.R
\name{predict_y}
\alias{predict_y}
\title{Returns predicted responses.}
\usage{
predict_y(model, testData = NULL)
}
\arguments{
\item{model}{a SplineTree object}

\item{testData}{The data to predict on. By default, uses the training set.}
}
\value{
A vector of predictions with rows corresponding to the testdata.
}
\description{
Returns predicted responses. Note that this function is most meaningful when used on spline tree objects that have an intercept.
}
\examples{
\dontrun{
split_formula <- BMI ~ HISP + WHITE + BLACK + SEX + Dad_Full_Work
  + Mom_Full_Work   + Age_first_weed + Age_first_smoke + Age_first_alc
  + Num_sibs + HGC_FATHER + HGC_MOTHER + Mag + News + Lib + Two_Adults_14
  + Mother_14 + Father_14 + STABLE_RESIDENCE + URBAN_14 + South_Birth
tree <- splineTree(split_formula, BMI~AGE, 'ID', nlsySample, degree=1,
  df=3, intercept=TRUE, cp=0.006, minNodeSize=20)
}
plot(predict_y(tree), tree$parms$data[[tree$parms$yvar]])
}
