% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_prediction.R
\name{predict_Y_tree}
\alias{predict_Y_tree}
\title{Predict responses for a testset using a single tree.}
\usage{
predict_Y_tree(tree, testset = tree$data)
}
\arguments{
\item{tree}{a splinetree object}

\item{testset}{the dataset to return predictions for. Defaults
to the dataaset used to build this tree.}
}
\value{
a matrix of predicted responses
}
\description{
Predict responses for a testset using a single tree.
}
