% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_evaluation.R
\name{predictCoeffs}
\alias{predictCoeffs}
\title{Predict spline coefficients from a splinetree object}
\usage{
predictCoeffs(tree, testset = tree$parms$flat_data)
}
\arguments{
\item{tree}{A splinetree object}

\item{testset}{The dataset to predict coefficients for. Default is the dataset used to make the tree.}
}
\value{
A matrix of spline coefficients. Dimension is number of units in test set by degrees of freedom of the spline.
}
\description{
Returns a matrix of spline coefficients for each observation in the testset. If no testset is provided,
returns predicted coefficients for the individuals in the dataset used to build the tree.
}
\details{
importFrom treeClust rpart.predict.leaves
}
\examples{
\dontrun{
split_formula <- BMI ~ HISP + WHITE + BLACK + SEX + Dad_Full_Work
  + Mom_Full_Work   + Age_first_weed + Age_first_smoke + Age_first_alc
  + Num_sibs + HGC_FATHER + HGC_MOTHER + Mag + News + Lib + Two_Adults_14
  + Mother_14 + Father_14 + STABLE_RESIDENCE + URBAN_14 + South_Birth
tree <- splineTree(split_formula, BMI~AGE, 'ID', nlsySample, degree=1,
  df=3, intercept=TRUE, cp=0.006, minNodeSize=20)
}
predictCoeffs(tree)
}
