% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_evaluation.R
\name{predictY}
\alias{predictY}
\title{Predictions from a splitted splinetree object}
\usage{
predictY(model, testData = NULL)
}
\arguments{
\item{model}{A splinetree object.}

\item{testData}{The data to return predictions for. If ommitted, uses the training data.}
}
\value{
A vector of predictions with rows corresponding to the testdata.
}
\description{
Returns a vector of predicted responses for the testData. If testData is ommitted,
returns predictions for the training data. This function is most meaningful if model$intercept==TRUE.
}
\examples{
\dontrun{
split_formula <-  ~ HISP + WHITE + BLACK + SEX + Dad_Full_Work
  + Mom_Full_Work   + Age_first_weed + Age_first_smoke + Age_first_alc
  + Num_sibs + HGC_FATHER + HGC_MOTHER + Mag + News + Lib + Two_Adults_14
  + Mother_14 + Father_14 + STABLE_RESIDENCE + URBAN_14 + South_Birth
tree <- splineTree(split_formula, BMI~AGE, 'ID', nlsySample, degree=1,
  df=3, intercept=TRUE, cp=0.006, minNodeSize=20)
}
plot(predictY(tree), tree$parms$data[[tree$parms$yvar]])
}
