% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_helpers.R
\name{pruneForest}
\alias{pruneForest}
\title{Prunes each tree in forest using a given complexity parameter.}
\usage{
pruneForest(forest, cp)
}
\arguments{
\item{forest}{A spline forest object}

\item{cp}{The complexity parameter that will be used to prune each tree (see rpart package documentation for detailed description of complexity parameter)}
}
\value{
A new splinetree forest object, where each tree has a new size
}
\description{
Prunes each tree in forest using a given complexity parameter.
}
\examples{
print(avSize(sample_forest))
print(avSize(pruneForest(sample_forest, cp=0.007)))
print(avSize(pruneForest(sample_forest, cp=0.01)))
}
