% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_evaluation.R
\name{R2_projected}
\alias{R2_projected}
\title{Computes an R^2-like measure that is based on the projected sum of squared errors.}
\usage{
R2_projected(model, includeIntercept = FALSE)
}
\arguments{
\item{model}{a splinetree tree object}

\item{includeIntercept}{If FALSE and if the model was built with an intercept, the projected squared errors are computed
while ignoring the intercept. If the model was built without an intercept, this parameter does not do anything.}
}
\description{
Computes an R^2-like measure that is based on the projected sum of squared errors. Can be used on trees whether or not they
were built with an intercept. If the tree was built with an intercept, there is the option to ignore the intercept in this
projection to isolate how well the tree clusters based on shape.
}
