% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_evaluation.R
\name{R2_y}
\alias{R2_y}
\title{Percent of variation in response explained by spline tree.}
\usage{
R2_y(model)
}
\arguments{
\item{model}{a splinetree tree object}
}
\value{
An R^2 goodness measure. 1-SSE/SST where SSE is the sum of squared errors between predicted responses and true
responses, and SST is sum of squared errors of true responses around population mean. Note that if the tree passed in was built
without an intercept, this function will return NULL.
}
\description{
Computes an R^2 measure for the spline tree with respect to prediction. Note that this metric is only meaningful if the spline tree object includes an intercept.
If the tree includes an intercept, the measure will be between 0 and 1.
}
\examples{
tree1 <- splineTree(BMI~HISP+WHITE+BLACK+HGC_MOTHER+SEX, BMI~AGE, "ID", nlsySample, degree=1, intercept=FALSE, cp=0.005)
R2_y(tree1)
}
