% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helper_fns.R
\name{rpartco}
\alias{rpartco}
\title{Calculates coordinates for tree plot}
\usage{
rpartco(tree, parms = paste(".rpart.parms", dev.cur(), sep = "."))
}
\arguments{
\item{tree}{a SplineTree object}

\item{parms}{a string}
}
\description{
Figures out the coordinates on the tree plot for the each mini trajectory plots.
Modified from code from the longRPart package.
}
\keyword{internal}
