% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_intercept_forest}
\alias{sample_intercept_forest}
\title{A pre-built random forest (with intercept) with 20 trees, built to the NLSY sample.}
\format{An object of class \code{list} of length 16.}
\usage{
sample_intercept_forest
}
\description{
The purpose of this pre-built forest is to demonstrate the forest evaluation functions without
needing to rebuild a forest (slow) every time. This forest uses an intercept.
}
\keyword{datasets}
