% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{spaghettiPlot}
\alias{spaghettiPlot}
\title{Create a facetted spaghetti plot of a splinetree model}
\usage{
spaghettiPlot(model, colors = NULL)
}
\arguments{
\item{model}{a splinetree object}

\item{colors}{optional arguement specifying colors to be used for each panel.}
}
\description{
Uses ggplot to create a paneled spaghetti plot of the data, where each panel corresponds to a terminal node in the tree.
Allows users to visualize homogeneity of trajectories within the terminal nodes of the tree.
}
\examples{
\dontrun{
split_formula <- BMI ~ HISP + WHITE + BLACK + SEX + Dad_Full_Work
  + Mom_Full_Work   + Age_first_weed + Age_first_smoke + Age_first_alc
  + Num_sibs + HGC_FATHER + HGC_MOTHER + Mag + News + Lib + Two_Adults_14
  + Mother_14 + Father_14 + STABLE_RESIDENCE + URBAN_14 + South_Birth
tree <- splineTree(split_formula, BMI~AGE, 'ID', nlsySample, degree=1,
  df=3, intercept=TRUE, cp=0.006, minNodeSize=20)
}
spaghettiPlot(tree)
}
