% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_split_forest.R
\name{splineforest_split}
\alias{splineforest_split}
\title{Custom rpart split function for spline random forests}
\usage{
splineforest_split(y, wt, x, parms = NULL, continuous)
}
\arguments{
\item{y}{the responses at this node}

\item{wt}{the weight of the responses}

\item{x}{the X data for this covariate}

\item{parms}{the basis matrix for the spline and the proportion of variables randomly sampled (diceProb)}

\item{continuous}{value is handled internally by rpart - tells us if this covariate is continuous or categorical (factor).}
}
\description{
Wrapper for split function required for the random forest functionality. This function is called once per covariate at each potential split.
Implements the random selection of variables; each variable is randomly selected to be included or excluded.
}
