% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_print.R
\name{stPrint}
\alias{stPrint}
\title{Print a spline tree object}
\usage{
stPrint(x, cp, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{The splinetree object}

\item{cp}{Optional- if provided, a pruned version of the tree will be printed. The tree will be
pruned using the provided cp as the complexity parameter.}

\item{digits}{Specifies how many digits of each coefficient should be printed}
}
\value{
A printout of the tree. The printout provides numbered labels for the terminal nodes,
a description of the split at each node, the number of observations found at each node, and the
predicted spline coefficients for each node.
}
\description{
Code adapted only slightly from the rpart base code for print.rpart to support the printing of
all coefficients.
}
