% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_print.R
\name{tree.depth}
\alias{tree.depth}
\title{Given a list of node numbers, returns the depth at which these appear in the tree.}
\usage{
tree.depth(nodes)
}
\description{
Used in printing and plotting.
Source: rpart
}
\keyword{internal}
