% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_print.R
\name{treeSummary}
\alias{treeSummary}
\title{Prints the tree frame.}
\usage{
treeSummary(model)
}
\arguments{
\item{model}{A splinetree object.}
}
\description{
Prints the tree frame.
}
\examples{
\dontrun{
split_formula <- BMI ~ HISP + WHITE + BLACK + SEX + Dad_Full_Work
  + Mom_Full_Work   + Age_first_weed + Age_first_smoke + Age_first_alc
  + Num_sibs + HGC_FATHER + HGC_MOTHER + Mag + News + Lib + Two_Adults_14
  + Mother_14 + Father_14 + STABLE_RESIDENCE + URBAN_14 + South_Birth
tree <- splineTree(split_formula, BMI~AGE, 'ID', nlsySample, degree=1,
  df=3, intercept=TRUE, cp=0.006, minNodeSize=20)
}
treeSummary(tree)
}
