% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Yimp}
\alias{Yimp}
\title{A pre-computed variable importance matrix for the example forest.}
\format{An object of class \code{matrix} with 10 rows and 3 columns.}
\usage{
Yimp
}
\description{
Once again, this is saved simply so that examples can be shown without recomputing this matrix every time.
The Y-variable-importance function is relatively inefficient.
}
\keyword{datasets}
