% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_helpers.R
\name{avSize}
\alias{avSize}
\title{Compute the average tree size in a forest}
\usage{
avSize(forest)
}
\arguments{
\item{forest}{A model returned by splineForest()}
}
\value{
The average number of terminal nodes in forest
}
\description{
Returns the average number of terminal nodes for trees in a forest
}
\examples{
avSize(forest)
}
