% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlsySample}
\alias{nlsySample}
\title{Baseline socioeconomic information and BMIs of 500 individuals.}
\format{A data frame with 8083 rows and 18 columns.
\describe{
  \item{ID}{jj}
  \item{News}{gj}
  \item{SEX}{gnegne}
  \item{AGE}{aa}
  \item{BLACK}{aa}
  \item{BMI}{aa}
  \item{Father_14}{aa}
  \item{HEIGHTIN}{aa}
  \item{HGC_FATHER}{aa}
  \item{HGC_MOTHER}{aa}
  \item{HISP}{aa}
  \item{Lib}{aa}
  \item{Mag}{aa}
  \item{Mother_14}{aa}
  \item{Num_sibs}{aa}
  \item{Two_Adults_14}{aa}
  \item{WEIGHT}{aa}
  \item{WHITE}{aa}

  ...
}}
\source{
\url{https://www.bls.gov/nls/nlsy79.htm}
}
\usage{
nlsySample
}
\description{
A dataset containing the body mass index (BMI) and baseline
socioeconomic information of 500 individuals from the National
Longitudinal Survey of Youth 1979 (NLSY), a freely available longitudinal dataset.
The 500 individuals were drawn randomly from among all NLSY respondents with at least 10
non-missing height/weight responses spread out over at least 20 years.
This dataset is used for many of our code examples.
}
\keyword{datasets}
