% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlsySample}
\alias{nlsySample}
\title{Baseline socioeconomic information and BMI of 100 individuals.}
\format{A data frame with 16126 rows and 34 columns.
\describe{
  \item{ID}{Unique identifier for each NLSY respondent}
  \item{News}{Did anyone in respondent's family subscribe to a newspaper when respondent was 14?}
  \item{SEX}{Respondent's sex. 1 denotes male, 2 denotes female.}
  \item{AGE}{Respondent's age}
  \item{BLACK}{Indicator for whether or not respondent's identified as Black}
  \item{BMI}{Respondent's body mass index - calculated from reported height and weight}
  \item{Father_14}{Was the respondent's father present when respondednt was 14}
  \item{HEIGHTIN}{Respondent's height}
  \item{HGC_FATHER}{Highest grade completed by respondent's father}
  \item{HGC_MOTHER}{Highest grade completed by respondent's mother}
  \item{HISP}{Indicator for whether or not respondent's race identified as Hispanic}
  \item{Lib}{Did anyone in respondent's family have a library card when respondent was 14?}
  \item{Mag}{Did anyone in respondent's family subscribe to a magazine when respondent was 14?}
  \item{Mother_14}{Was respondent's mother present when respondent was 14?}
  \item{Num_sibs}{Number of siblings of respondent}
  \item{Two_Adults_14}{Were two adults present in respondent's household at age 14?}
  \item{WEIGHT}{Respondent's weight}
  \item{WHITE}{Indicator for whether or not respondent identified as white.}
  \item{HGC}{Highest grade completed by respondent}
  \item{POVSTAT}{Indicator denoting whether or not respondent is in poverty}
  \item{Dad_Full_Work}{Did respondent's father work full time (>35 hours) at time of first interview? NA if father not on record. }
  \item{Mom_Full_Work}{Did respondent's mother work full time (>35 hours) at time of first interview? NA if mother not on record.}
  \item{Age_first_weed}{Age that respondent reported first using marijuana. If they reported never using marijuana, recorded as 100.}
  \item{Age_first_smoke}{Age that respondent reported first using tobacco. If they reported never using tobacco, recorded as 100.}
  \item{Live_with_parents}{Did respondent report living in same household as the parents on record? }
  \item{Age_first_alc}{Age that respondent reported first drinking alcohol. If they reported never drinking alcohol, recorded as 100.}
  \item{Mom_dad_household}{Did the respondent's mother and father live in same household at first interview?}
  \item{Father_on_record}{Did the respondent have a father listed in the survey record at first interview?}
  \item{Mother_on_record}{Did the respondent have a mother listed in the survey record at first interview?}
  \item{STABLE_RESIDENCE}{Did the respondent live in the same house from birth to age 14?}
  \item{URBAN_14}{Did the respondent live in an urban area at age 14? 1 denotes town or city, 2 denotes country not farm, 3 denotes farm}
  \item{RACE}{Race, as recorded by NLSY. 1 denotes Hispanic, 2 denotes Black, 3 denotes White.}
  \item{South_Birth}{Indicator for whether or not respondent was born in the south (south as defined
   by NLSY).}
  \item{HOUSEHOLD_14}{Household identifier for survey respondent}
}}
\source{
\url{https://www.bls.gov/nls/nlsy79.htm}
}
\usage{
nlsySample
}
\description{
A dataset containing the body mass index (BMI) and baseline
socioeconomic information of 100 individuals from the National
Longitudinal Survey of Youth 1979 (NLSY), a freely available longitudinal dataset.
The 1000 individuals were drawn randomly from among all NLSY respondents with at least 10
non-missing height/weight responses spread out over at least 20 years.
This dataset is used in the package vignettes and code examples. Only a small
subset of the variables available from the NLSY are included here.
See https://www.bls.gov/nls/nlsy79.htm for more
}
\keyword{internal}
