% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_evaluation.R
\name{predict_y}
\alias{predict_y}
\title{Returns predicted responses.}
\usage{
predict_y(model, testData = model$parms$data)
}
\arguments{
\item{model}{a SplineTree object}

\item{testData}{The data to predict on. By default, uses the training set.}
}
\value{
A vector of predictions with rows corresponding to the testdata.
}
\description{
Returns predicted responses. Note that this function is most meaningful when used on spline tree objects that have an intercept.
}
\examples{
tree1 <- splineTree(BMI~HISP+WHITE+BLACK+HGC_MOTHER+SEX, BMI~AGE, "ID", nlsySample, degree=1, intercept=FALSE, cp=0.005)
plot(predict_y(tree1), tree1$parms$data[[tree1$parms$yvar]])
}
