% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_evaluation.R
\name{predict_y_training}
\alias{predict_y_training}
\title{Predict responses for the training data}
\usage{
predict_y_training(model)
}
\arguments{
\item{model}{a splinetree object}
}
\value{
A vector of predicted responses
}
\description{
Returns a vector of predicted responses for the dataset used to build the tree
}
\details{
Calling predictY(model) and predict_y_training(model) return identical results, because when no test data is
provided to predictY(), the default is to use the training set. This is a slightly faster version that
can be used when you know that you wish to predict on the training data.
}
\keyword{internal}
