% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_evaluation.R
\name{predict_y_training}
\alias{predict_y_training}
\title{Predict responses for the training data}
\usage{
predict_y_training(model)
}
\arguments{
\item{model}{a model created with splineTree()}
}
\value{
A vector of predicted responses where each element in the vector corresponds to a row in model$parms$data.
}
\description{
Calling predictY(model) and predict_y_training(model) return identical results, because when no test data is
provided to predictY(), the default is to use the training set. This is a slightly faster version that
can be used when you know that you wish to predict on the training data. It is faster because it takes advantage
of the relationship between model$parms$flat_data and model$parms$data.
}
\keyword{internal}
