% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_prediction.R
\name{predictYForest}
\alias{predictYForest}
\title{Predict responses using random forest.}
\usage{
predictYForest(forest, method = "oob", testdata = NULL)
}
\arguments{
\item{forest}{a splinetree forest object}

\item{method}{a string; either "oob", "itb", or "all". "oob" is the default value.
if "oob", predictions for a given data point are made only using trees for which this
data point was "out of the bag" (not in the bootstrap sample). If "itb", predictions for
a given data point are made using onle the trees for which this datapoint was in the bag (in the bootstrap sample).
If "all", all trees are used.}

\item{testdata}{the test data to make preditions for. If this is provided, then
all trees are used for all datapoints.}
}
\value{
a matrix of predicted responses.
}
\description{
Uses the forest to make predictions of responses for individuals at given times
in the training sample or on a new sample. If the testdata parameter is null,
then predictions are given on the training sample according to one of three possible
methods. The supplied method parameter must be either "oob", "itb", or "all". Note that this method
should only be used on trees that have an intercept. Otherwise, the Y predictions will
not be accurate at all.
}
