% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{spaghettiPlot}
\alias{spaghettiPlot}
\title{Create a facetted spaghetti plot of a splinetree model}
\usage{
spaghettiPlot(model, colors = NULL)
}
\arguments{
\item{model}{a splinetree object}

\item{colors}{optional arguement specifying colors to be used for each panel.}
}
\description{
Uses ggplot to create a paneled spaghetti plot of the data, where each panel corresponds to a terminal node in the tree.
Allows users to visualize homogeneity of trajectories within the terminal nodes of the tree.
}
\examples{
tree <- splineTree(BMI~HISP+WHITE+BLACK+HGC_MOTHER+SEX, BMI~AGE, "ID", nlsySample,
degree=2, intercept=TRUE, cp=0.005)
spaghettiPlot(tree)
}
