% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplineForest.R
\name{splineForest}
\alias{splineForest}
\title{Build a splinetree random forest}
\usage{
splineForest(splitFormula, tformula, idvar, data, knots = NULL, df = NULL,
  degree = 3, intercept = FALSE, nGrid = 7, ntree = 50,
  diceProb = 0.3, cp = 0.001)
}
\arguments{
\item{splitFormula}{Formula specifying the longitudinal response variable and the time-constant variables that will be used for splitting in the tree.}

\item{tformula}{Formula specifying the longitudinal response variable and the variable that acts as the time variable.}

\item{idvar}{The name of the variable that serves as the ID variable for grouping observations. Must be in quotes}

\item{data}{dataframe that contains all variables specified in the formulas- in long format.}

\item{knots}{Specified locations for internal knots in the spline basis. Defaults to NULL, which corresponds to no internal knots.}

\item{df}{Degrees of freedom of the spline basis. If this is specified but the knots parameter is NULL, then the appropriate number of internal knots
will be added at quantiles of the training data. If both df and knots are unspecified, the spline basis will have no internal knots.}

\item{degree}{Specifies degree of spline basis used in the tree.}

\item{intercept}{Specifies whether or not the splitting process will consider the intercept coefficient of the spline projections.
Defaults to FALSE, which means that the tree will split based on trajectory shape, ignoring response level.}

\item{nGrid}{Number of grid points to evaluate projection sum of squares at. The default is 7, which corresponds to evaluating projections
at the endpoints and quintiles of the time variable.}

\item{ntree}{Number of trees in the forest}

\item{diceProb}{Probability of selecting a variable to included as a candidate for each split.}

\item{cp}{Complexity parameter passed to the rpart building process.}

\item{minNodeSize}{Minimum number of observational units that can be in a terminal node. Controls tree size and helps avoid overfitting.}
}
\value{
A splineforest object, which stores a list of splinetree objects (in model$Trees), along with information about the
spline basis used (model$intercept, model$innerKnots, model$boundaryKnots, etc.), and information about which datapoints were
used to build each tree (model$oob_indices and model$index).
}
\description{
Builds an ensemble of regression trees for longitudinal or functional data using the spline projection method. The resulting object
is a list of splinetree objects along with some additional information. All parameters are used in the same way that they are used in the splineTree
function. The additional parameter ntree specifies how many trees should be in the ensemble, and diceProb controls the
probability of selecting a given variable for split consideration at a node.
}
\details{
The ensemble method is highly similar to the random forest methodology of Brieman (2001). Each tree in the ensemble is fit to a bootstrap sample
of the data. At each node of each tree, only a subset of the split variables are considered candidates for the split. In our methodology,
the subset of variables considered at each node is determined by a random process. The diceProb parameter specifies the probability that a given variable
will be selected at a certain node. Because the method is based on probability, the same number of variables are not considered for splitting at each node
(as in the randomForest package). Note that if diceProb is small and the number of variables in the splitFormula is also small, there is a high probability that
no variables will be considered for splitting at a certain node, which is problematic. The fewer total variables there are, the larger diceProb should be to
ensure good results.
}
\examples{
splitForm <-BMI~HISP+WHITE+BLACK+HGC_MOTHER+HGC_FATHER+SEX+Num_sibs
model1 <- splineForest(splitForm, BMI~AGE, 'ID', nlsySample, degree=1, intercept=FALSE, cp=0.005, ntree=30)
}
