% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplineTree.R
\name{splineTree}
\alias{splineTree}
\title{Build a splinetree object}
\usage{
splineTree(splitFormula, tformula, idvar, data, knots = NULL, df = NULL,
  degree = 3, intercept = FALSE, nGrid = 7, gridPoints = NULL,
  minNodeSize = 10, cp = 0.01)
}
\arguments{
\item{splitFormula}{Formula specifying the longitudinal response variable and the time-constant variables that will be used for splitting in the tree.}

\item{tformula}{Formula specifying the longitudinal response variable and the variable that acts as the time variable.}

\item{idvar}{The name of the variable that serves as the ID variable for grouping observations. Must be in quotes}

\item{data}{dataframe that contains all variables specified in the formulas- in long format.}

\item{knots}{Specified locations for internal knots in the spline basis. Defaults to NULL, which corresponds to no internal knots.}

\item{df}{Degrees of freedom of the spline basis. If this is specified but the knots parameter is NULL, then the appropriate number of internal knots
will be added at quantiles of the training data. If both df and knots are unspecified, the spline basis will have no internal knots.}

\item{degree}{Specifies degree of spline basis used in the tree.}

\item{intercept}{Specifies whether or not the splitting process will consider the intercept coefficient of the spline projections.
Defaults to FALSE, which means that the tree will split based on trajectory shape, ignoring response level.}

\item{nGrid}{Number of grid points to evaluate projection sum of squares at. If gridPoints
is not supploed, this arguement will be used and the projection sum of squares will be evaluated
at quantiles of the time variable.}

\item{gridPoints}{Optional. A vector of numbers that will be used as the grid on which to evaluate the
projection sum of squares. Should fall roughly within the range of the time variable.}

\item{minNodeSize}{Minimum number of observational units that can be in a terminal node. Controls tree size and helps avoid overfitting.}

\item{cp}{Complexity parameter passed to the rpart building process.}
}
\value{
An rpart object with additional splinetree-specific information stored in model$parms.
}
\description{
Builds a regression tree for longitudinal or functional data using the spline projection method. The underlying tree building process uses the rpart package,
and the splinetree object is an rpart object with additional stored information. The parmaters df, knots, degree, interept, and nGrid allow for flexibility
in the spline basis used for splitting. The parameters minNodeSize and cp allow for flexibility in controlling the size of the final tree.
}
\examples{
splitForm <-BMI~HISP+WHITE+BLACK+HGC_MOTHER+HGC_FATHER+SEX+Num_sibs
model1 <- splineTree(splitForm, BMI~AGE, 'ID', nlsySample, degree=1, intercept=FALSE, cp=0.005)
model2 <- splineTree(splitForm, BMI~AGE, 'ID', nlsySample, degree=3, intercept=TRUE, cp=0.005)
stPrint(model1)
treeSummary(model1)
stPlot(model1)
stPlot(model2)
R2_projected(model1)
R2_projected(model2)
}
