% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{splineTreePlot}
\alias{splineTreePlot}
\title{Tree plot of a spline tree}
\usage{
splineTreePlot(model, colors = NULL)
}
\arguments{
\item{model}{a splinetree object}

\item{colors}{a list of colors that will be used for the terminal nodes (if NULL, will use a rainbow)}
}
\description{
Creates a tree plot of a SplineTree object.
}
