% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{stPlot}
\alias{stPlot}
\title{Plots a Spline Tree, showing the tree and the trajectories for comparison.}
\usage{
stPlot(model, colors = NULL)
}
\arguments{
\item{model}{a SplineTree object}

\item{colors}{a list of colors that will be used for the trajectories (if NULL, will automatically select colors from
rainbow color scheme.}
}
\description{
Creates a two paneled plot of a SplineTree object that shows both the tree and the trajectories side by side.
}
\examples{
form <- BMI~HISP+WHITE+BLACK+HGC_MOTHER+SEX
model <- splineTree(form, BMI~AGE, 'ID', nlsySample, degree=1, intercept=FALSE, cp=0.007)
stPlot(model)
}
