% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_print.R
\name{stPrint}
\alias{stPrint}
\title{Print a spline tree object}
\usage{
stPrint(tree, cp, digits = getOption("digits"))
}
\arguments{
\item{tree}{A splinetree object}

\item{cp}{Optional- if provided, a pruned version of the tree will be printed. The tree will be
pruned using the provided cp as the complexity parameter.}

\item{digits}{Specifies how many digits of each coefficient should be printed}
}
\value{
A printout of the tree. The printout provides numbered labels for the terminal nodes,
a description of the split at each node, the number of observations found at each node, and the
predicted spline coefficients for each node.
}
\description{
Code adapted only slightly from the rpart base code for print.rpart to support the printing of
all coefficients.
}
\examples{
\dontrun{
split_formula <- BMI ~ HISP + WHITE + BLACK + SEX + Dad_Full_Work
  + Mom_Full_Work   + Age_first_weed + Age_first_smoke + Age_first_alc
  + Num_sibs + HGC_FATHER + HGC_MOTHER + Mag + News + Lib + Two_Adults_14
  + Mother_14 + Father_14 + STABLE_RESIDENCE + URBAN_14 + South_Birth
 tree <- splineTree(split_formula, BMI~AGE, 'ID', nlsySample, degree=1,
   df=3, intercept=TRUE, cp=0.006, minNodeSize=20)
}
stPrint(tree)
}
