% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_useful_fns.R
\name{terminalNodeSummary}
\alias{terminalNodeSummary}
\title{Returns a printed summary of the paths to the terminal nodes in the tree.}
\usage{
terminalNodeSummary(tree, node = NULL)
}
\arguments{
\item{tree}{a splinetree object}
}
\value{
a list, indexed by terminal node numbers, where each
element of the list holds the path to the terminal node.
}
\description{
Importantly, provides the node numbers for each node that are found in the 'where'
attribute of the tree. Note that these nubmers differ from those shown in the rpart
printout of the tree.
}
