% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_useful_fns.R
\name{terminalNodeSummary}
\alias{terminalNodeSummary}
\title{Prints a summary of a terminal node in a tree}
\usage{
terminalNodeSummary(tree, node = NULL)
}
\arguments{
\item{tree}{A splinetree object}

\item{node}{The number of the node that you want summarized. To see which nodes correspond to
which numbers, see stPrint(tree).}
}
\description{
If no arguement is provided for the parameter \code{node}, summaries are printed for every
terminal node. Otherwise, the summary of just the requested node is printed.
}
