% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_useful_fns.R
\name{treeSize}
\alias{treeSize}
\title{Returns number of terminal nodes in a tree.}
\usage{
treeSize(model)
}
\arguments{
\item{model}{A splinetree object, or any rpart object}
}
\value{
Number of terminal nodes in tree
}
\description{
Returns number of terminal nodes in a tree.
}
