% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_useful_fns.R
\name{treeSize}
\alias{treeSize}
\title{Returns number of terminal nodes in a tree.}
\usage{
treeSize(model)
}
\arguments{
\item{model}{A splinetree object, or any rpart object}
}
\value{
Number of terminal nodes in tree
}
\description{
Returns number of terminal nodes in a tree.
}
\examples{
splitForm <-BMI~HISP+WHITE+BLACK+HGC_MOTHER+HGC_FATHER+SEX+Num_sibs
model1 <- splineTree(splitForm, BMI~AGE, 'ID', nlsySample, degree=1, intercept=FALSE, cp=0.005)
treeSize(model1)
}
